+++ crates/ruma-client-api/src/relations.rs
@@ -1,0 +1,3 @@
+//! Endpoints for querying relations
+
+pub mod get_events_relating_to_event;